move generateCode to shared package
